Fixed conversion warnings

+++ CrazyCanvas/Source/World/LevelManager.cpp
@@ -80,7 +80,7 @@ bool LevelManager::Init()
 							//Load module and store byte representation
 							auto moduleByteRepresentationIt = moduleByteRepresentations.find(moduleDesc.Filename);
 
-							uint32 currentByteRepresentationSize = byteRepresentation.size();
+							uint32 currentByteRepresentationSize = (uint32)byteRepresentation.size();
 							if (moduleByteRepresentationIt == moduleByteRepresentations.end())
 							{
 								byte* pData;
@@ -112,7 +112,7 @@ bool LevelManager::Init()
 							moduleDesc.Translation.y = translation.HasMember("y") ? translation["y"].GetFloat() : 0.0f;
 							moduleDesc.Translation.z = translation.HasMember("z") ? translation["z"].GetFloat() : 0.0f;
 
-							uint32 currentByteRepresentationSize = byteRepresentation.size();
+							uint32 currentByteRepresentationSize = (uint32)byteRepresentation.size();
 							byteRepresentation.resize(currentByteRepresentationSize + sizeof(moduleDesc.Translation));
 							memcpy(byteRepresentation.data() + currentByteRepresentationSize, glm::value_ptr(moduleDesc.Translation), sizeof(moduleDesc.Translation));
 						}
